yawar-api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
supabase
python-jose[cryptography]
pydantic
python-dotenv
